+++ src/mol-canvas3d/passes/tracing.ts
@@ -43,7 +43,6 @@ export const TracingParams = {
     rendersPerFrame: PD.Interval([1, 16], { min: 1, max: 64, step: 1 }, { description: 'Number of rays per pixel each frame. May be adjusted to reach targetFps but will stay within given interval.' }),
     targetFps: PD.Numeric(30, { min: 0, max: 120, step: 0.1 }, { description: 'Target FPS per frame. If observed FPS is lower or higher, some parameters may get adjusted.' }),
     steps: PD.Numeric(32, { min: 1, max: 1024, step: 1 }),
-    firstStepSize: PD.Numeric(0.01, { min: 0.001, max: 1, step: 0.001 }),
     refineSteps: PD.Numeric(4, { min: 0, max: 8, step: 1 }, { description: 'Number of refine steps per ray hit. May be lower to reach targetFps.' }),
     rayDistance: PD.Numeric(256, { min: 1, max: 8192, step: 1 }, { description: 'Maximum distance a ray can travel (in world units).' }),
     thicknessMode: PD.Select('auto', PD.arrayToOptions(['auto', 'fixed'] as const)),
@@ -371,10 +370,6 @@ export class TracingPass {
             ValueCell.update(this.traceRenderable.values.dSteps, steps);
             needsUpdateTrace = true;
         }
-        if (this.traceRenderable.values.dFirstStepSize.ref.value !== props.firstStepSize) {
-            ValueCell.update(this.traceRenderable.values.dFirstStepSize, props.firstStepSize);
-            needsUpdateTrace = true;
-        }
         if (this.traceRenderable.values.dRefineSteps.ref.value !== refineSteps) {
             ValueCell.update(this.traceRenderable.values.dRefineSteps, refineSteps);
             needsUpdateTrace = true;
@@ -440,7 +435,6 @@ const TraceSchema = {
     dGlow: DefineSpec('boolean'),
     dBounces: DefineSpec('number'),
     dSteps: DefineSpec('number'),
-    dFirstStepSize: DefineSpec('number'),
     dRefineSteps: DefineSpec('number'),
     uRayDistance: UniformSpec('f'),
 
@@ -490,7 +484,6 @@ function getTraceRenderable(ctx: WebGLContext, colorTexture: Texture, normalText
         dGlow: ValueCell.create(true),
         dBounces: ValueCell.create(4),
         dSteps: ValueCell.create(32),
-        dFirstStepSize: ValueCell.create(0.01),
         dRefineSteps: ValueCell.create(4),
         uRayDistance: ValueCell.create(256),